Removed old paddle
