+++ front-end/app/pong/scene/OneForAllScene.tsx
@@ -12,9 +12,10 @@ import EndModal from "@/components/Pong/EndModal";
 import { CornerBorder } from "@/components/Pong/Border";
 import { PongGameEvents } from "@/components/Pong/PongGameEvents";
 import { PongSocketEvents } from "@/components/Pong/PongSocketEvents";
-import { PaddleControl } from "../hooks/Paddle";
+import { PaddleControl } from "../../../components/Pong/Paddle";
 
 // FIXME: Scoring doesnt work anymore
+
 /**
  * The OneForAllScene component is a Three.js scene representing a 4 player Pong game that includes various elements such as paddles,
  * ball, borders, camera, countdown, scoreboard, and a modal for displaying the winner.
